Fixed handling of overflow

+++ OnStep.ino
@@ -558,8 +558,8 @@ void loop2() {
   if (!isSlewing()) nv.poll();
   
   // WORKLOAD MONITORING -------------------------------------------------------------------------------
-  long this_loop_micros=micros();
-  loop_time=this_loop_micros-last_loop_micros;
+  unsigned long this_loop_micros=micros();
+  loop_time=(long)(this_loop_micros-last_loop_micros);
   if (loop_time > worst_loop_time) worst_loop_time=loop_time;
   last_loop_micros=this_loop_micros;
   average_loop_time=(average_loop_time*49+loop_time)/50;
